`js2r-rename-var': don't go to start of symbol
Previously, `js2r-rename-var' would go to the start of the symbol being renamed
first. This makes inserting edits at `point' inconvenient, because one has to
navigate back.

Solve this by first computing a `point'-delta (how far `point' is from the start
of the symbol), and use that to position the other cursors. Don't `goto-char' to
the start of the symbol for the main cursor.

Fixes #122.